refactor(dashboard): split AgeBadge into its own file and use max-w-xs
Move AgeBadge out of summary-card.tsx so summary-card.tsx only contains
SummaryCard. Replace the inline-block wrapper on the liquidity screen's
total balance card with max-w-xs for a cleaner constraint.
