+++ src/components/dashboard/age-badge.tsx
@@ -1,0 +1,19 @@
+import { useEffect, useState } from 'react';
+
+function formatAge(ms: number): string {
+  const s = Math.max(0, Math.floor(ms / 1000));
+  if (s < 60) return `${s}s ago`;
+  const m = Math.floor(s / 60);
+  if (m < 60) return `${m}m ${s % 60}s ago`;
+  const h = Math.floor(m / 60);
+  return `${h}h ${m % 60}m ago`;
+}
+
+export function AgeBadge({ timestamp }: { timestamp?: string }): JSX.Element {
+  const [now, setNow] = useState(Date.now());
+  useEffect(() => {
+    const id = setInterval(() => setNow(Date.now()), 1000);
+    return () => clearInterval(id);
+  }, []);
+  return <>{timestamp ? formatAge(now - new Date(timestamp).getTime()) : '-'}</>;
+}
