+++ src/screens/dashboard-financial-overview.screen.tsx
@@ -2,7 +2,8 @@ import { useSessionContext } from '@dfx.swiss/react';
 import { SpinnerSize, StyledLoadingSpinner } from '@dfx.swiss/react-components';
 import { useEffect, useMemo, useState } from 'react';
 import { BalanceBarChart } from 'src/components/dashboard/latest-balance-bar-chart';
-import { AgeBadge, SummaryCard } from 'src/components/dashboard/summary-card';
+import { AgeBadge } from 'src/components/dashboard/age-badge';
+import { SummaryCard } from 'src/components/dashboard/summary-card';
 import { TotalBalanceLongChart } from 'src/components/dashboard/total-balance-long-chart';
 import { FinancialLogEntry, LatestBalanceResponse } from 'src/dto/dashboard.dto';
 import { useDashboard } from 'src/hooks/dashboard.hook';
